+++ src/helper.cpp
@@ -31,7 +31,7 @@ void vector_to_file(std::vector<double> const &vec, std::string const &filename)
     file.close();
 }
 
-std::array<float, 6> decompose_matrix(std::array<float, 6> mat) {
+std::array<double, 6> decompose_matrix(std::array<double, 6> mat) {
 
     /*
      * Matrix looks like:
@@ -39,41 +39,41 @@ std::array<float, 6> decompose_matrix(std::array<float, 6> mat) {
      * d e ty       m3 m4 m5
      */
 
-    float a = mat[0];
-    float b = mat[1];
-    float d = mat[3];
-    float e = mat[4];
-    float tx = mat[2];
-    float ty = mat[5];
+    double a = mat[0];
+    double b = mat[1];
+    double d = mat[3];
+    double e = mat[4];
+    double tx = mat[2];
+    double ty = mat[5];
 
-    float sx = sqrt(a * a + d * d);
-    float sy = sqrt(b * b + e * e);
+    double sx = sqrt(a * a + d * d);
+    double sy = sqrt(b * b + e * e);
 
-    float phi = atan2(d, a);
+    double phi = atan2(d, a);
     if (phi < 0.1)
         phi += 2. * M_PI;
 
-    float shear = atan2(-b, e) - phi;
+    double shear = atan2(-b, e) - phi;
     if (shear < -6.1)
         shear += 2. * M_PI;
 
-    std::array<float, 6> result = {sx, sy, shear, phi, tx, ty};
+    std::array<double, 6> result = {sx, sy, shear, phi, tx, ty};
     // return <sx, sy, shear, rot, tx ,ty>
     return result;
 }
 
-std::array<float, 6> compose_matrix(std::vector<float> parameters) {
-    float sx = parameters[0];
-    float sy = parameters[1];
-    float shear = parameters[2];
-    float rot = parameters[3];
+std::array<double, 6> compose_matrix(std::vector<double> parameters) {
+    double sx = parameters[0];
+    double sy = parameters[1];
+    double shear = parameters[2];
+    double rot = parameters[3];
 
-    std::array<float, 6> m = {
-            sx * (float) cos(rot),
-            -sy * (float) sin(rot + shear),
+    std::array<double, 6> m = {
+            sx * cos(rot),
+            -sy * sin(rot + shear),
             parameters[4],
-            sx * (float) sin(rot),
-            sy * (float) cos(rot + shear),
+            sx * sin(rot),
+            sy * cos(rot + shear),
             parameters[5],
     };
     return m;
